pyproject.toml
[project]
name = "mockendpoint-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
]
